Put stats files in a separate directory.
Import the stats files that will be used for the elo estimates.
