+++ fishtest/fishtest/views.py
@@ -18,7 +18,7 @@ from pyramid.security import remember, forget, authenticated_userid, has_permiss
 from pyramid.view import view_config, forbidden_view_config
 from pyramid.httpexceptions import HTTPFound, HTTPBadRequest
 
-import stat_util
+import fishtest.stats.stat_util
 
 # For caching the tests output
 cache_time = 2
@@ -631,19 +631,19 @@ def format_results(run_results, run):
     sprt = run['args']['sprt']
     state = sprt.get('state', '')
 
-    stats = stat_util.SPRT(run_results,
-                           elo0=sprt['elo0'],
-                           alpha=sprt['alpha'],
-                           elo1=sprt['elo1'],
-                           beta=sprt['beta'],
-                           drawelo=sprt['drawelo'])
+    stats = fishtest.stats.stat_util.SPRT(run_results,
+                                          elo0=sprt['elo0'],
+                                          alpha=sprt['alpha'],
+                                          elo1=sprt['elo1'],
+                                          beta=sprt['beta'],
+                                          drawelo=sprt['drawelo'])
     result['llr'] = stats['llr']
     result['info'].append('LLR: %.2f (%.2lf,%.2lf) [%.2f,%.2f]' % (stats['llr'], stats['lower_bound'], stats['upper_bound'], sprt['elo0'], sprt['elo1']))
   else:
     if 'pentanomial' in run_results.keys():
-      elo, elo95, los = stat_util.get_elo(run_results['pentanomial'])
+      elo, elo95, los = fishtest.stats.stat_util.get_elo(run_results['pentanomial'])
     else:
-      elo, elo95, los = stat_util.get_elo([WLD[1],WLD[2],WLD[0]])
+      elo, elo95, los = fishtest.stats.stat_util.get_elo([WLD[1],WLD[2],WLD[0]])
 
     # Display the results
     eloInfo = 'ELO: %.2f +-%.1f (95%%)' % (elo, elo95)
